+++ CHANGELOG.md
@@ -9,6 +9,12 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Added
+
+- Added missing `urlv2` field to `DatasetStorage` data model.
+- Added missing `constraints` field to `Session` data model.
+- Added missing `description` field to `Group` data model.
+
 ### Removed
 
 - Removed deprecated `url` field from `FileInfo` data model.